requirements.txt
# Core
pytest
python-dotenv
pandas
requests
numpy<2.0.0
black
pylint

# RAG & NLP
langchain
langchain-community
langchain-mistralai
faiss-cpu
mistralai
sentence-transformers
tiktoken

# API
fastapi
uvicorn
httpx

# Evaluation
ragas